+++ CHANGELOG.md
@@ -6,7 +6,7 @@
     types. It also works for callbacks that are Swift properties.
 - Bug fixes:
   + Fixed documentation comments handling for Java interfaces and Swift protocols.
-  + Fixed a Cpp-Swift-Cpp object round-trip segfault issue.
+  + Fixed several Cpp-Swift-Cpp object round-trip segfault issues.
   + Naming a parameter-less method "release" does not produce un-compilable Swift code anymore.
 
 ## 3.15.0